Create class of cpi, loco and permutation
