+++ src/hidimstat/__init__.py
@@ -10,9 +10,8 @@ from .permutation_test import permutation_test, permutation_test_pval
 from .scenario import multivariate_1D_simulation
 from .empirical_thresholding import empirical_thresholding
 from .stat_tools import zscore_from_pval
-from .cpi import CPI
-from .loco import LOCO
-from .permutation_importance import permutation_importance
+from .permutation_importance import permutation_importance, cpi, loco
+from .permutation_importance_classes import PermutationImportance, CPI, LOCO
 
 try:
     from ._version import __version__
@@ -39,5 +38,8 @@ __all__ = [
     "zscore_from_pval",
     "CPI",
     "LOCO",
+    "PermutationImportance",
+    "cpi",
+    "loco",
     "permutation_importance",
 ]
